+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/warning/WarningObjectBuilder.java
@@ -1,0 +1,26 @@
+
+package com.commercetools.api.models.warning;
+
+import java.util.*;
+
+import io.vrap.rmf.base.client.utils.Generated;
+
+/**
+ * WarningObjectBuilder
+ */
+@Generated(value = "io.vrap.rmf.codegen.rendering.CoreCodeGenerator", comments = "https://github.com/commercetools/rmf-codegen")
+public class WarningObjectBuilder {
+
+    public com.commercetools.api.models.warning.ImageProcessingOngoingWarningBuilder imageProcessingOngoingBuilder() {
+        return com.commercetools.api.models.warning.ImageProcessingOngoingWarningBuilder.of();
+    }
+
+    /**
+     * factory method for an instance of WarningObjectBuilder
+     * @return builder
+     */
+    public static WarningObjectBuilder of() {
+        return new WarningObjectBuilder();
+    }
+
+}