+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/warning/WarningObjectImpl.java
@@ -1,0 +1,93 @@
+
+package com.commercetools.api.models.warning;
+
+import java.time.*;
+import java.util.*;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.*;
+
+import io.vrap.rmf.base.client.ModelBase;
+import io.vrap.rmf.base.client.utils.Generated;
+
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
+import org.apache.commons.lang3.builder.ToStringBuilder;
+import org.apache.commons.lang3.builder.ToStringStyle;
+
+/**
+ *  <p>Represents a warning related to the returned response.</p>
+ */
+@Generated(value = "io.vrap.rmf.codegen.rendering.CoreCodeGenerator", comments = "https://github.com/commercetools/rmf-codegen")
+public class WarningObjectImpl implements WarningObject, ModelBase {
+
+    private String code;
+
+    private String message;
+
+    /**
+     * create instance with all properties
+     */
+    @JsonCreator
+    WarningObjectImpl(@JsonProperty("code") final String code, @JsonProperty("message") final String message) {
+        this.code = code;
+        this.message = message;
+    }
+
+    /**
+     * create empty instance
+     */
+    public WarningObjectImpl() {
+    }
+
+    /**
+     *  <p>Identifier for the type of warning.</p>
+     */
+
+    public String getCode() {
+        return this.code;
+    }
+
+    /**
+     *  <p>Contains information about the returned response.</p>
+     */
+
+    public String getMessage() {
+        return this.message;
+    }
+
+    public void setMessage(final String message) {
+        this.message = message;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o)
+            return true;
+
+        if (o == null || getClass() != o.getClass())
+            return false;
+
+        WarningObjectImpl that = (WarningObjectImpl) o;
+
+        return new EqualsBuilder().append(code, that.code)
+                .append(message, that.message)
+                .append(code, that.code)
+                .append(message, that.message)
+                .isEquals();
+    }
+
+    @Override
+    public int hashCode() {
+        return new HashCodeBuilder(17, 37).append(code).append(message).toHashCode();
+    }
+
+    @Override
+    public String toString() {
+        return new ToStringBuilder(this, ToStringStyle.SHORT_PREFIX_STYLE).append("code", code)
+                .append("message", message)
+                .build();
+    }
+
+}