+++ commercetools/commercetools-sdk-java-api/src/main/java-predicates-generated/com/commercetools/api/predicates/query/product/ProductQueryBuilderDsl.java
@@ -114,4 +114,18 @@ public class ProductQueryBuilderDsl {
             p -> new CombinationQueryPredicate<>(p, ProductQueryBuilderDsl::of));
     }
 
+    public CombinationQueryPredicate<ProductQueryBuilderDsl> warnings(
+            Function<com.commercetools.api.predicates.query.warning.WarningObjectQueryBuilderDsl, CombinationQueryPredicate<com.commercetools.api.predicates.query.warning.WarningObjectQueryBuilderDsl>> fn) {
+        return new CombinationQueryPredicate<>(
+            ContainerQueryPredicate.of()
+                    .parent(ConstantQueryPredicate.of().constant("warnings"))
+                    .inner(fn.apply(com.commercetools.api.predicates.query.warning.WarningObjectQueryBuilderDsl.of())),
+            ProductQueryBuilderDsl::of);
+    }
+
+    public CollectionPredicateBuilder<ProductQueryBuilderDsl> warnings() {
+        return new CollectionPredicateBuilder<>(BinaryQueryPredicate.of().left(new ConstantQueryPredicate("warnings")),
+            p -> new CombinationQueryPredicate<>(p, ProductQueryBuilderDsl::of));
+    }
+
 }