+++ changes.md
@@ -3,7 +3,15 @@
 <details>
 <summary>Added Property(s)</summary>
 
-- added property `staged` to type `ProductVariantDeletedMessage`
-- added property `staged` to type `ProductVariantDeletedMessagePayload`
+- added property `warnings` to type `ProductTailoring`
+- added property `warnings` to type `Product`
+</details>
+
+
+<details>
+<summary>Added Type(s)</summary>
+
+- added type `ImageProcessingOngoingWarning`
+- added type `WarningObject`
 </details>